server/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==3.0.0
Flask-Login==0.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7  # PostgreSQL database adapter